requirements.txt
pyqrcode
requests
colorama
python-socketio
